Remove watchdog, replace with a simple background thread
